+++ lib/sidekiq/delay_extensions/action_mailer.rb
@@ -12,12 +12,14 @@ module Sidekiq
     #    UserMailer.delay.send_welcome_email(new_user)
     #    UserMailer.delay_for(5.days).send_welcome_email(new_user)
     #    UserMailer.delay_until(5.days.from_now).send_welcome_email(new_user)
-    class DelayedMailer
-      include Sidekiq::Job
-
-      def perform(yml)
-        (target, method_name, args) = ::Sidekiq::DelayExtensions::YAML.unsafe_load(yml)
-        msg = target.public_send(method_name, *args)
+    class DelayedMailer < GenericJob
+      def _perform(target, method_name, *args, **kwargs)
+        msg =
+          if kwargs.empty?
+            target.public_send(method_name, *args)
+          else
+            target.public_send(method_name, *args, **kwargs)
+          end
         # The email method can return nil, which causes ActionMailer to return
         # an undeliverable empty message.
         if msg
@@ -29,16 +31,24 @@ module Sidekiq
     end
 
     module ActionMailer
+      def sidekiq_delay_proxy
+        if Sidekiq::DelayExtensions.use_generic_proxy
+          GenericProxy
+        else
+          Proxy
+        end
+      end
+
       def sidekiq_delay(options = {})
-        Proxy.new(DelayedMailer, self, options)
+        sidekiq_delay_proxy.new(DelayedMailer, self, options)
       end
 
       def sidekiq_delay_for(interval, options = {})
-        Proxy.new(DelayedMailer, self, options.merge("at" => Time.now.to_f + interval.to_f))
+        sidekiq_delay_proxy.new(DelayedMailer, self, options.merge("at" => Time.now.to_f + interval.to_f))
       end
 
       def sidekiq_delay_until(timestamp, options = {})
-        Proxy.new(DelayedMailer, self, options.merge("at" => timestamp.to_f))
+        sidekiq_delay_proxy.new(DelayedMailer, self, options.merge("at" => timestamp.to_f))
       end
       alias_method :delay, :sidekiq_delay
       alias_method :delay_for, :sidekiq_delay_for