+++ lib/sidekiq/delay_extensions/active_record.rb
@@ -14,26 +14,28 @@ module Sidekiq
     # Please note, this is not recommended as this will serialize the entire
     # object to Redis.  Your Sidekiq jobs should pass IDs, not entire instances.
     # This is here for backwards compatibility with Delayed::Job only.
-    class DelayedModel
-      include Sidekiq::Job
-
-      def perform(yml)
-        (target, method_name, args) = ::Sidekiq::DelayExtensions::YAML.unsafe_load(yml)
-        target.__send__(method_name, *args)
-      end
+    class DelayedModel < GenericJob
     end
 
     module ActiveRecord
+      def sidekiq_delay_proxy
+        if Sidekiq::DelayExtensions.use_generic_proxy
+          GenericProxy
+        else
+          Proxy
+        end
+      end
+
       def sidekiq_delay(options = {})
-        Proxy.new(DelayedModel, self, options)
+        sidekiq_delay_proxy.new(DelayedModel, self, options)
       end
 
       def sidekiq_delay_for(interval, options = {})
-        Proxy.new(DelayedModel, self, options.merge("at" => Time.now.to_f + interval.to_f))
+        sidekiq_delay_proxy.new(DelayedModel, self, options.merge("at" => Time.now.to_f + interval.to_f))
       end
 
       def sidekiq_delay_until(timestamp, options = {})
-        Proxy.new(DelayedModel, self, options.merge("at" => timestamp.to_f))
+        sidekiq_delay_proxy.new(DelayedModel, self, options.merge("at" => timestamp.to_f))
       end
       alias_method :delay, :sidekiq_delay
       alias_method :delay_for, :sidekiq_delay_for