+++ lib/sidekiq/delay_extensions/generic_proxy.rb
@@ -6,6 +6,9 @@ module Sidekiq
   module DelayExtensions
     SIZE_LIMIT = 8_192
 
+    singleton_class.attr_accessor :use_generic_proxy
+    self.use_generic_proxy = false
+
     class Proxy < BasicObject
       def initialize(performable, target, options = {})
         @performable = performable
@@ -27,6 +30,60 @@ module Sidekiq
         @performable.client_push({"class" => @performable,
                                   "args" => [marshalled],
                                   "display_class" => "#{@target}.#{name}"}.merge(@opts))
+      end
+    end
+
+    class GenericProxy < BasicObject
+      def initialize(performable, target, options = {})
+        @performable = performable
+        @target = target
+        @opts = options
+      end
+
+      def method_missing(name, *args, **kwargs) # rubocop:disable Style/MissingRespondToMissing
+        begin
+          has_no_kwargs = kwargs.nil? || kwargs.empty?
+          valid_json_args =
+            if has_no_kwargs
+              # if args.last.is_a?(Hash) && args.last.keys.any? { |key| key.is_a?(Symbol) }
+              #   kwargs = args.pop.symbolize_keys
+              [*args]
+            else
+              [*args, ::JSON.parse(::JSON.dump(kwargs))]
+            end
+          obj = [@target.name, name&.to_s, valid_json_args]
+          marshalled = ::JSON.dump(obj)
+        rescue ::JSON::ParserError
+          obj = if kwargs&.any?
+            [@target.name, name&.to_s, *args, **kwargs]
+          else
+            [@target.name, name&.to_s, *args]
+          end
+          ::Sidekiq.logger.warn {
+            "Non-JSON args passed to Sidekiq delayed job. obj=#{obj.inspect}"
+          }
+          marshalled = ::YAML.dump(obj)
+        end
+        # Debug notes:
+        # marshalled = ::YAML.dump(valid_json_args)
+        # marshalled = ::YAML.to_json(valid_json_args)
+        # value = {
+        #   target_name: @target.name,
+        #   name: name,
+        #   args: args,
+        #   kwargs: kwargs,
+        #   valid_json_args: valid_json_args,
+        #   obj: obj,
+        #   marshalled: marshalled,
+        # }
+        # ::STDOUT.puts value.inspect
+        if marshalled.size > SIZE_LIMIT
+          ::Sidekiq.logger.warn { "#{@target}.#{name} job argument is #{marshalled.bytesize} bytes, you should refactor it to reduce the size" }
+        end
+        valid_opts = @opts.stringify_keys
+        @performable.client_push({"class" => @performable,
+                                  "args" => [marshalled],
+                                  "display_class" => "#{@target}.#{name}"}.merge(valid_opts))
       end
     end
   end